### 2024-06-15 #6894 @editorialbot
[REVIEW]: Cthulhu: An Open Source Molecular and Atomic Cross Section Computation Code for Substellar Atmospheres
**Submitting author:** @arnav-agrawal (<a href="http://orcid.org/0000-0003-2944-0600">Arnav Agrawal</a>)
**Repository:** https://github.com/MartianColonist/Cthulhu
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
