### 2024-06-25 #6894 @LorenzoMugnai
Hi @arnav-agrawal,

I have been reviewing the documentation and found it to be a bit too minimalist. If I may, here are a few suggestions for improvement:

- **Installation Instructions**: I noticed that you recommend installing the software from source via GitHub. While this is fine, I wonder why you suggest using the pip editable mode (-e). Could you please explain the necessity of this in the documentation?
- **Functionality Documentation**: Your API documentation appears to be built using autoapi (as mentioned on the API page). Could you add some explanatory text around it, at least in the index, to help users navigate the documentation more effectively?
- **Community Guidelines**: I found these either missing or not easily accessible. Could you please add a dedicated section for community guidelines in the documentation?
- **minors**: Regarding the home page, I suggest including a link to the "Introduction" section and to the paper to be cited (unless the paper is the JOSS one and is not published yet, of course).

Thank you for your attention to these suggestions.
