### 2024-06-26 #6894 @LorenzoMugnai
Hi again,

Regarding the paper, I found it quite enjoyable to read. I have a few comments, mostly aimed at making it more accessible for non-expert users, whom you identify as the preferred audience for the didactic purpose of your code:

- You start with fundamental assumptions in the field (e.g., "Without accurate cross sections, one cannot reliably measure the chemical composition of substellar atmospheres."), which is a nice way to introduce the framework. However, I feel the paper could benefit from additional references or a few more lines to support such statements.
- You mention that the code can run on a CPU laptop for "smaller line lists." What do you mean by "smaller"? Can you provide benchmarks, such as reference computation times and memory usage?
- You state that the HDF5 output can be used with retrieval codes. Could you provide examples of codes natively supported by your output format (such as Poseidon, Taurex, etc.)? Or does the user need to adapt your output for these codes? Including references or citations for some retrieval codes would be helpful.
- You graphically show that the cross sections produced can be used to generate transmission spectra. Can you explain how this can be done for a non-expert user, perhaps by referencing the integration with modelling codes?
- I feel the statement of need could be more effective if moved just after the summary, before delving into the details of Excalibur's computational process. Have you considered this?
- You close the manuscript by inviting the reader to provide suggestions. It would be useful to reference your "contribution guidelines," encouraging users to join the discussion on a dedicated channel (perhaps GitHub?).
