### 2024-06-26 #6894 @LorenzoMugnai
Another note (sorry for the several messages),

I see that there are automated tests in your repository. However, they don't seem to be linked to any CI, GitHub Action, or coverage tool. Have you considered including a GitHub Action for CI with a dedicated badge? [Here](https://github.com/MartianColonist/POSEIDON/blob/main/.github/workflows/pytest_testing.yml) is an example for POSEIDON, the retrieval code developed by your co-author and hosted in the same repository as Excalibur.

Thank you!
