### 2024-06-28 #6894 @arjunsavel
@dfm hope you're well! Just wanted to double-check that I'm not COI-d out of reviewing. I'm on a paper (maybe a few?) with the second author, but it's part of a rather large collaboration (JWST ERS, and the limb asymmetry sub-unit). I don't think I've had much scientific conversation with the second author.
