### 2024-06-29 #6894 @dfm
@arjunsavel — Thanks for checking in and disclosing this potential COI. To me, this sounds like a small enough conflict that I'm happy to waive it and continue this review. If anyone involved (reviewer or author) has any concerns, please feel free to comment here or email me directly. Thanks all!
