### 2024-07-19 #6894 @MartianColonist
@dfm We're considering a rebranding of this code due to a name clash with the Caltech exoplanet pipeline EXCALIBUR (https://excalibur.ipac.caltech.edu/). Will the JOSS review break if we change the name of our package?
