### 2024-07-20 #6894 @dfm
@MartianColonist — There shouldn't be any issues with renaming. Just let us know here what it ends up as and I'll track down any technical issues that arise!
