### 2024-08-07 #6894 @arnav-agrawal
Hi @dfm @LorenzoMugnai @arjunsavel,

Thanks again for all your comments. We've addressed all the code-related things, and I'll quote some replies to each of you below this with the specific concerns you brought up.

So as a general comment, we have rebranded the name of the package to 'Cthulhu' instead of 'Excalibur'. This is the new repository link: [](https://github.com/MartianColonist/Cthulhu)

We've updated internal ReadTheDocs links and such to update this renaming. Everything is still the same, just the name changed. Hopefully nothing breaks in the JOSS pipeline! We're currently updating the paper/some figures accordingly, and I'll post a comment in the next few days when that is done.
