### 2024-08-07 #6894 @arnav-agrawal
> Thanks to the authors for putting together a nicely designed package that is very useful, both scientifically and pedagogically. I’ll likely be a user myself in the coming months!
>
> I have some minor comments below. Some I would prefer to see addressed before recommending acceptance, whereas other needn’t necessarily be addressed (I’ll indicate which are the latter).
>
> ### Installation (not necessary for acceptance)
> * I had some dependency issues installing on my local machine (all were fine when installing in a fresh google colab notebook). In particular, it seems as though the trapz module isn’t available in all scipy versions. Perhaps consider pinning some package versions (scipy, numpy especially), and / or providing an environment.yml file.
> * I’d also suggest moving to a pyproject.toml file instead of setup.py for a more modern installation flow.
>
> ### Testing
> * I attempted to run the tests (aside from the atomic line tests), and I wasn’t able to import them (the tests expect excalibur, while the package is installed as Excalibur).
> * Some tests fail. Namely:
>
>   * test_get_default_linelist in exomol
>   * test_create_id_dict in hitran
>   * test_det_broad() in exomol
>
> ### Misc.
> * In line with Lorenzo, I suggest creating a contributing.md file.
>
> ### Documentation (not necessary for acceptance, as I believe the Quickstart and intro are sufficiently thorough)
> * Home page
>
>   * transition — > transition
>
> * Introduction
>
>   * Put Hargreaves and gordon in parentheses?
>   * NIST also in parentheses?
>   * Gordon in parentheses?
>   * Could replace the transmission-specific example with “when light is absorbed or emitted in an exoplanet atmosphere…”
> * API
>
>   * Docstrings aren’t complete in:
>
>     * excalibur.plot
>     * hdf5_package
>     * Constants?
>     * Calculate
>     * Broadening
>     * Voigt
> * Really cool to have the cluster documentation!
> * Some of the tutorials are incomplete.
>
>   * Such as the temperature--pressure one
>
>     * Didn’t work, couldn’t find T=3000 file
>
>       * Error line: nu3, sigma3  = read_cross_section_file(species = species, database = database, filename = 'CO_T3000.0K_log_P0.0_H2-He_sigma.txt'
>     * Ls: didn’t find the rounded version
>
>       * Only found CO_T3000K_log_P0.0_H2-He_sigma.txt
>   * Comparing isotopologues
>
> ### Paper comments
> * Rapidly-evolving —> Rapidly evolving
> * Could even motivate opacity’s importance more by citing how sensitive high-resolution spectroscopy can be to line list differences (e.g., Brogi and Line 2019).
> * Don’t need ELTs in parentheses, because the phrase is not used again.
> * Cite JWST and some of the ELT papers?
> * Please cite a few theory papers re: _requiring_ cross-sections.
> * Is it possible to make the legend text larger in Figure 1?
> * Perhaps one sentence about what Excalibur does under the hood to calculate the cross-sections would be helpful.
> * A HDF5 file --> an HDF5 file
> * A few more references would be helpful in laying out the radiative transfer. Cite the slant optical depth, for instance one of the earlier Fortney papers?
> * Please the higher-quality spectra of JWST.
> * higher fidelity --> higher-fidelity
> * Please some brown dwarf modeling codes.
> * Please cite some high-resolution ground-based datasets
> * Beginner’s —> beginners
> * Speed-dependent —> speed-dependent
>
> Please let me know if you have any questions about my comments!

@arjunsavel Thanks for the suggestions!

- We are updating the setup.py to catch some of the version mismatches that you found. There's some deprecated functions (like the trapz one you found) that we hadn't caught earlier.
- For now, we'll stick with the setup.py framework we have, as it works consistently. Thank you for suggesting the pyproject.toml framework, we hadn't heard about it. It's definitely something we'll consider moving to in the future.
- The tests all work now, and we have a CI/CD pipeline in place for running them on new pushes to the dev or main branch.

We are still working on updating the paper and will let you know once that's complete, thank you for your detailed suggestions.
