### 2024-08-09 e047599
Minor paper wording edits
