+++ joss/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly-evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) and the upcoming Extremely Large Telescopes (ELTs), are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere. Without accurate cross sections, one cannot reliably measure the chemical composition of substellar atmospheres. 
+Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) and the upcoming Extremely Large Telescopes, are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere. Without accurate cross sections, one cannot reliably measure the chemical composition of substellar atmospheres. 
 
 Cross sections are typically pre-computed on a grid of pressures and temperatures from large databases of quantum mechanical transitions (line lists), such as ExoMol [@Tennyson:2020], HITRAN [@Gordon:2022], HITEMP [@Rothman:2010], and VALD [@Pakhomov:2017]. However, the process of calculating cross sections from line lists is often computationally demanding and has required complex and specialized tools. We aim here to lower the access barrier for users to learn how to calculate molecular and atomic cross sections. 
 
@@ -47,7 +47,7 @@ The purpose of the `Cthulhu` package is schematically represented in \autoref{fi
 
 The first use of `Cthulhu` is to download existing molecular line lists from online databases. `Cthulhu`'s `summon` function can automatically download lines lists from ExoMol and HITRAN/HITEMP (for the latter a user must make an account on [https://hitran.org/](https://hitran.org/)) and reformat the line lists into space-efficient HDF5 files. Ancillary input files required to calculate cross sections, such as partition functions and pressure broadening files, are also downloaded automatically. Alternatively, the user may manually download a line list from their respective websites and point `Cthulhu` to the directory hosting the files. VALD line lists must be downloaded manually by a user with an account on [http://vald.astro.uu.se/](http://vald.astro.uu.se/) (given the terms of use for VALD3), but we provide instructions on how to do this in the `Cthulhu` documentation. `Cthulhu` currently supports ExoMol, HITRAN, HITEMP, and VALD line lists, though we welcome user requests for additional line list databases support. Once a line list has been downloaded, the user can move onto the next major use case of `Cthulhu`, computing cross sections.
 
[…]
