+++ joss/paper.md
@@ -17,7 +17,7 @@ authors:
     orcid: 0000-0003-4816-3469
     affiliation: "3, 4, 2"
 affiliations:
-  - name: Johns Hopkins Applied Physics Laboratory, 11100 Johns Hopkins Rd., Laurel, MD 20723, USA
+  - name: Johns Hopkins University Applied Physics Laboratory, 11100 Johns Hopkins Rd., Laurel, MD 20723, USA
     index: 1
   - name: Department of Astronomy and Carl Sagan Institute, Cornell University, 122 Sciences Drive, Ithaca, NY 14853, USA
     index: 2
@@ -33,25 +33,11 @@ bibliography: paper.bib
 
 # Summary
 
-Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) and the upcoming Extremely Large Telescopes, are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere. Without accurate cross sections, one cannot reliably measure the chemical composition of substellar atmospheres. 
+Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) and the upcoming Extremely Large Telescopes, are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere. One cannot reliably calculate spectra of substellar atmospheres without accurate cross sections, which can subsequently influence measurements of chemical abundances and other atmospheric properties [e.g., @Hedges:2016; @Gharib-Nezhad:2019; @Anisman:2022].
 
-Cross sections are typically pre-computed on a grid of pressures and temperatures from large databases of quantum mechanical transitions (line lists), such as ExoMol [@Tennyson:2020], HITRAN [@Gordon:2022], HITEMP [@Rothman:2010], and VALD [@Pakhomov:2017]. However, the process of calculating cross sections from line lists is often computationally demanding and has required complex and specialized tools. We aim here to lower the access barrier for users to learn how to calculate molecular and atomic cross sections. 
+Cross sections are typically pre-computed on a grid of pressures and temperatures from large databases of quantum mechanical transitions (line lists), such as ExoMol [@Tennyson:2020], HITRAN [@Gordon:2022], HITEMP [@Rothman:2010], and VALD [@Pakhomov:2017]. However, the process of calculating cross sections from line lists is often computationally demanding and has required complex and specialised tools. We aim here to lower the access barrier for users to learn how to calculate molecular and atomic cross sections. 
 
-`Cthulhu` is a fully Python package that rapidly calculates cross sections from atomic and molecular line lists. `Cthulhu` includes modules to automatically download molecular line lists from online databases and compute cross sections on a user-specified temperature, pressure, and wavenumber grid. `Cthulhu` requires only CPUs and can run on a user's laptop (for smaller line lists) or on a large cluster in parallel (for billions of lines). `Cthulhu` includes in-depth Jupyter tutorials in the online documentation. Finally, `Cthulhu` is intended not only for research purposes, but as an educational tool to demystify the process of making cross sections for atmospheric models.
-
-# Computing Molecular and Atomic Cross Sections with `Cthulhu`
-
[…]
