+++ joss/paper.bib
@@ -133,6 +133,132 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Hedges:2016,
+       author = {{Hedges}, Christina and {Madhusudhan}, Nikku},
+        title = "{Effect of pressure broadening on molecular absorption cross sections in exoplanetary atmospheres}",
+      journal = {\mnras},
+     keywords = {methods: laboratory, planets and satellites: atmospheres, planetary systems, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2016,
+        month = may,
+       volume = {458},
+       number = {2},
+        pages = {1427-1449},
+          doi = {10.1093/mnras/stw278},
+archivePrefix = {arXiv},
+       eprint = {1602.00751},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016MNRAS.458.1427H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Gharib-Nezhad:2019,
+       author = {{Gharib-Nezhad}, Ehsan and {Line}, Michael R.},
+        title = "{The Influence of H$_{2}$O Pressure Broadening in High-metallicity Exoplanet Atmospheres}",
+      journal = {\apj},
+     keywords = {molecular data, planets and satellites: atmospheres, planets and satellites: composition, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2019,
+        month = feb,
+       volume = {872},
+       number = {1},
+          eid = {27},
+        pages = {27},
+          doi = {10.3847/1538-4357/aafb7b},
+archivePrefix = {arXiv},
+       eprint = {1809.02548},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ApJ...872...27G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Anisman:2022,
+       author = {{Anisman}, Lara O. and {Chubb}, Katy L. and {Changeat}, Quentin and {Edwards}, Billy and {Yurchenko}, Sergei N. and {Tennyson}, Jonathan and {Tinetti}, Giovanna},
+        title = "{Cross-sections for heavy atmospheres: H$_{2}$O self-broadening}",
+      journal = {\jqsrt},
+     keywords = {Exoplanet atmospheres, Water vapor, Opacities, Radiative transfer, Line broadening, Super-Earths, mini-Neptunes, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2022,
+        month = jun,
+       volume = {283},
+          eid = {108146},
+        pages = {108146},
+          doi = {10.1016/j.jqsrt.2022.108146},
+archivePrefix = {arXiv},
+       eprint = {2203.02335},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022JQSRT.28308146A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{MacDonald:2023,
+       author = {{MacDonald}, Ryan J.},
+        title = "{POSEIDON: A Multidimensional Atmospheric Retrieval Code for Exoplanet Spectra}",
+      journal = {The Journal of Open Source Software},
+     keywords = {Exoplanets, Exoplanet atmospheres, Atmospheric retrieval, Bayesian statistics, Python, Open source software, JWST, Spectroscopy},
+         year = 2023,
+        month = jan,
+       volume = {8},
[…]
