+++ joss/paper.bib
@@ -136,7 +136,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Hedges:2016,
        author = {{Hedges}, Christina and {Madhusudhan}, Nikku},
         title = "{Effect of pressure broadening on molecular absorption cross sections in exoplanetary atmospheres}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: laboratory, planets and satellites: atmospheres, planetary systems, Astrophysics - Earth and Planetary Astrophysics},
          year = 2016,
         month = may,
@@ -154,7 +154,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Gharib-Nezhad:2019,
        author = {{Gharib-Nezhad}, Ehsan and {Line}, Michael R.},
         title = "{The Influence of H$_{2}$O Pressure Broadening in High-metallicity Exoplanet Atmospheres}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {molecular data, planets and satellites: atmospheres, planets and satellites: composition, Astrophysics - Earth and Planetary Astrophysics},
          year = 2019,
         month = feb,
@@ -173,7 +173,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Anisman:2022,
        author = {{Anisman}, Lara O. and {Chubb}, Katy L. and {Changeat}, Quentin and {Edwards}, Billy and {Yurchenko}, Sergei N. and {Tennyson}, Jonathan and {Tinetti}, Giovanna},
         title = "{Cross-sections for heavy atmospheres: H$_{2}$O self-broadening}",
-      journal = {\jqsrt},
+      journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
      keywords = {Exoplanet atmospheres, Water vapor, Opacities, Radiative transfer, Line broadening, Super-Earths, mini-Neptunes, Astrophysics - Earth and Planetary Astrophysics},
          year = 2022,
         month = jun,
@@ -185,6 +185,25 @@ archivePrefix = {arXiv},
        eprint = {2203.02335},
  primaryClass = {astro-ph.EP},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2022JQSRT.28308146A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{MacDonald:2022,
+       author = {{MacDonald}, Ryan J. and {Lewis}, Nikole K.},
+        title = "{TRIDENT: A Rapid 3D Radiative-transfer Model for Exoplanet Transmission Spectra}",
+      journal = {Astrophysical Journal},
+     keywords = {Exoplanet atmospheres, Exoplanet atmospheric composition, Hot Jupiters, Transmission spectroscopy, Spectroscopy, Exoplanets, Radiative transfer, 487, 2021, 753, 2133, 1558, 498, 1335, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2022,
+        month = apr,
+       volume = {929},
+       number = {1},
+          eid = {20},
+        pages = {20},
+          doi = {10.3847/1538-4357/ac47fe},
+archivePrefix = {arXiv},
+       eprint = {2111.05862},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJ...929...20M},
[…]
