+++ joss/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) [@Gardner:2023] and the upcoming Extremely Large Telescopes [@Skidmore:2015, @Fanson:2022, @Padovani:2023], are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere [@Seager:2010, @Heng:2017]. Reliably calculating spectra of substellar atmospheres requires accurate cross sections, without which measurements of chemical abundances and other atmospheric properties can be biased [e.g., @Hedges:2016; @Gharib-Nezhad:2019; @Anisman:2022]. 
+Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) [@Gardner:2023] and the upcoming Extremely Large Telescopes [@Skidmore:2015; @Fanson:2022; @Padovani:2023], are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere [@Seager:2010, @Heng:2017]. Reliably calculating spectra of substellar atmospheres requires accurate cross sections, without which measurements of chemical abundances and other atmospheric properties can be biased [e.g., @Anisman:2022; @Hedges:2016; @Gharib-Nezhad:2019]. 
 
 Cross sections are typically pre-computed on a grid of pressures and temperatures from large databases of quantum mechanical transitions (line lists), such as ExoMol [@Tennyson:2020], HITRAN [@Gordon:2022], HITEMP [@Rothman:2010], and VALD [@Pakhomov:2017]. However, the process of calculating cross sections from line lists is often computationally demanding and has required complex and specialised tools. We aim here to lower the access barrier for users to learn how to calculate molecular and atomic cross sections. 
 
@@ -41,7 +41,7 @@ Cross sections are typically pre-computed on a grid of pressures and temperature
 
 # Statement of Need
 
-JWST has recently significantly expanded the number of exoplanet and brown dwarfs with high-quality spectra spanning a wide wavelength range [e.g., @Miles:2023, @Beiler:2023, @Carter:2024, @Welbanks:2024]. High-fidelity spectra motivate detailed intercomparisons of exoplanet and brown dwarf modeling codes [e.g. @Barstow:2020], which often identify opacity database differences as a key modelling limitation. Ground-based high spectral resolution datasets [e.g. @Snellen:2010, @Birkby:2017, @Pelletier:2023] also critically rely on up-to-date opacity data, since older inaccurate line lists can lead to non-detections of molecules via cross correlations [e.g., @Merritt:2020] and propagate into retrieved chemical abundances [e.g., @Brogi:2019]. However, despite the key need to continually refine exoplanet and brown dwarf models with the latest state-of-the-art opacities, the process of calculating molecular and atomic cross sections is a non-trivial task that is typically outside the speciality of many exoplanet and brown dwarf researchers.
[…]
