+++ joss/paper.bib
@@ -419,7 +419,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Fortney2005,
+@ARTICLE{Fortney:2005,
        author = {{Fortney}, Jonathan J.},
         title = "{The effect of condensates on the characterization of transiting planet atmospheres with transmission spectroscopy}",
       journal = {Monthly Notices of the Royal Astronomical Society},
@@ -437,7 +437,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@PHDTHESIS{MacDonald2019,
+@PHDTHESIS{MacDonald:2019,
        author = {{MacDonald}, Ryan John},
         title = "{Revealing the nature of exoplanetary atmospheres}",
        school = {University of Cambridge, UK},
