+++ joss/paper.md
@@ -33,17 +33,17 @@ bibliography: paper.bib
 
 # Summary
 
-Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) [@Gardner:2023] and the upcoming Extremely Large Telescopes [@Skidmore:2015; @Fanson:2022; @Padovani:2023], are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere [@Seager:2010, @Heng:2017]. Reliably calculating spectra of substellar atmospheres requires accurate cross sections, without which measurements of chemical abundances and other atmospheric properties can be biased [e.g., @Anisman:2022; @Hedges:2016; @Gharib-Nezhad:2019]. 
+Atmospheric studies of exoplanets and brown dwarfs are a cutting-edge and rapidly evolving area of astrophysics research. Powerful new telescopes, such as the James Webb Space Telescope (JWST) [@Gardner:2023] and the upcoming Extremely Large Telescopes [@Skidmore:2015; @Fanson:2022; @Padovani:2023], are able to capture in detail spectra of planets and brown dwarfs and thereby probe their chemical composition and physical properties. Calculating models of exoplanet or brown dwarf spectra requires knowledge of the wavelength-dependent absorption of light (cross sections) by the molecules and atoms in the atmosphere [@Seager:2010; @Heng:2017]. Reliably calculating spectra of substellar atmospheres requires accurate cross sections, without which measurements of chemical abundances and other atmospheric properties can be biased [e.g., @Anisman:2022; @Hedges:2016; @Gharib-Nezhad:2019]. 
 
-Cross sections are typically pre-computed on a grid of pressures and temperatures from large databases of quantum mechanical transitions (line lists), such as ExoMol [@Tennyson:2020], HITRAN [@Gordon:2022], HITEMP [@Rothman:2010], and VALD [@Pakhomov:2017]. However, the process of calculating cross sections from line lists is often computationally demanding and has required complex and specialised tools. We aim here to lower the access barrier for users to learn how to calculate molecular and atomic cross sections. 
+Cross sections are typically pre-computed on a grid of pressures and temperatures from large databases of quantum mechanical transitions (line lists), such as ExoMol [@Tennyson:2020], HITRAN [@Gordon:2022], HITEMP [@Rothman:2010], and VALD [@Pakhomov:2017]. However, calculating cross sections from line lists is often computationally demanding and has required complex and specialised tools. We aim here to lower the access barrier for users to learn how to calculate molecular and atomic cross sections. 
 
 `Cthulhu` is a fully Python package that rapidly calculates cross sections from atomic and molecular line lists. `Cthulhu` includes modules to automatically download molecular line lists from online databases and compute cross sections on a user-specified temperature, pressure, and wavenumber grid. `Cthulhu` requires only CPUs and can run on a user's laptop (for smaller line lists with < 100 million lines) or on a large cluster in parallel (for many billion lines). `Cthulhu` includes in-depth Jupyter tutorials in the online documentation. Finally, `Cthulhu` is intended not only for research purposes, but as an educational tool to demystify the process of making cross sections for atmospheric models.
 
 # Statement of Need
 
[…]
