+++ joss/paper.bib
@@ -189,7 +189,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{MacDonald:2022,
-       author = {{MacDonald}, Ryan J. and {Lewis}, Nikole K.},
+       author = {{MacDonald} and {Lewis}},
         title = "{TRIDENT: A Rapid 3D Radiative-transfer Model for Exoplanet Transmission Spectra}",
       journal = {Astrophysical Journal},
      keywords = {Exoplanet atmospheres, Exoplanet atmospheric composition, Hot Jupiters, Transmission spectroscopy, Spectroscopy, Exoplanets, Radiative transfer, 487, 2021, 753, 2133, 1558, 498, 1335, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
@@ -208,7 +208,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{MacDonald:2023,
-       author = {{MacDonald}, Ryan J.},
+       author = {{MacDonald}},
         title = "{POSEIDON: A Multidimensional Atmospheric Retrieval Code for Exoplanet Spectra}",
       journal = {The Journal of Open Source Software},
      keywords = {Exoplanets, Exoplanet atmospheres, Atmospheric retrieval, Bayesian statistics, Python, Open source software, JWST, Spectroscopy},
@@ -438,7 +438,7 @@ archivePrefix = {arXiv},
 }
 
 @PHDTHESIS{MacDonald:2019,
-       author = {{MacDonald}, Ryan John},
+       author = {{MacDonald}},
         title = "{Revealing the nature of exoplanetary atmospheres}",
        school = {University of Cambridge, UK},
          year = 2019,
