+++ joss/paper.md
@@ -53,7 +53,7 @@ The purpose of the `Cthulhu` package is schematically represented in \autoref{fi
 
 The first use of `Cthulhu` is to download existing molecular line lists from online databases. `Cthulhu`'s `summon` function can automatically download lines lists from ExoMol and HITRAN/HITEMP and reformat the line lists into space-efficient HDF5 files. Ancillary input files required to calculate cross sections, such as partition functions and pressure broadening files, are also downloaded automatically. Alternatively, the user may manually download a line list from their respective websites and point `Cthulhu` to the directory hosting the files. VALD line lists must be downloaded manually by a user with an account on [http://vald.astro.uu.se/](http://vald.astro.uu.se/) (given the terms of use for VALD3), but we provide instructions on how to do this in the `Cthulhu` documentation. `Cthulhu` currently supports ExoMol, HITRAN, HITEMP, and VALD line lists, though we welcome user requests for additional line list databases support. Once a line list has been downloaded, the user can move onto the next major use case of `Cthulhu`, computing cross sections.
 
-The foremost feature of `Cthulhu` is its ability to straightforwardly compute atomic and molecular cross sections at high speeds (typically > 100,000 lines per second on one CPU). `Cthulhu` calculates cross sections via a generalisation of the Vectorised Voigt method [@Yurchenko:2018], whereby complex derivatives are used to perturb a grid of pre-computed template Voigt profiles to the specific properties of each given line [see @MacDonald:2019 for the mathematical description of the underlying algorithm]. `Cthulhu` is accessible, as it does not require GPUs, can run on a standard laptop, and as a fully Python code it is easy for beginners to install and use. To compute a cross section, a user simply calls `Cthulhu`'s `compute_cross_section` function, specifying the location of the line list, the temperature and pressure, and the wavenumber range. More advanced users can specify custom settings via optional arguments (e.g. Voigt wing cutoffs, intensity cutoffs, or a user-provided pressure broadening file). The documentation and function doc strings explain the various arguments users can provide to `compute_cross_section`. The computed cross section is output by default as a .txt file in the `output` folder on the user's machine, which can be readily used to create an opacity database in a user-preferred format for a specific radiative transfer code. `Cthulhu` also offers utility functions to combine multiple cross section .txt files (e.g. a grid of cross sections for different temperature and pressures for one or more chemical species) into an HDF5 cross section database, as illustrated in our [`quick start`](https://cthulhu.readthedocs.io/en/latest/content/quick_start.html) guide.
[…]
