### 2024-08-23 c04e831
Update JOSS paper Figure
