+++ joss/paper.md
@@ -49,7 +49,7 @@ We have built `Cthulhu` to provide a user-friendly tool for beginners to learn h
 
 The purpose of the `Cthulhu` package is schematically represented in \autoref{fig:Cthulhu_architecture}. Here we walk through this flowchart, highlighting major use cases of `Cthulhu` and the package's role in the broader process of modelling exoplanetary and brown dwarf atmospheres.
 
-![The role and applications of the `Cthulhu` Python package. `Cthulhu` can download molecular and atomic line lists and calculate the corresponding absorption cross sections as a function of temperature, pressure, and wavenumber. Cross sections made by `Cthulhu` can be used in radiative transfer codes to calculate model spectra of exoplanet and brown dwarf atmospheres. \label{fig:Cthulhu_architecture}](figures/Cthulhu_JOSS_Figure.png){width=100%}
+![The role and applications of the `Cthulhu` Python package. `Cthulhu` can download molecular and atomic line lists and calculate the corresponding absorption cross sections as a function of temperature, pressure, and wavenumber. Cross sections made by `Cthulhu` can be used in radiative transfer codes to calculate model spectra of exoplanet and brown dwarf atmospheres. \label{fig:Cthulhu_architecture}](figures/Cthulhu_Figure.png){width=100%}
 
 The first use of `Cthulhu` is to download existing molecular line lists from online databases. `Cthulhu`'s `summon` function can automatically download lines lists from ExoMol and HITRAN/HITEMP and reformat the line lists into space-efficient HDF5 files. Ancillary input files required to calculate cross sections, such as partition functions and pressure broadening files, are also downloaded automatically. Alternatively, the user may manually download a line list from their respective websites and point `Cthulhu` to the directory hosting the files. VALD line lists must be downloaded manually by a user with an account on [http://vald.astro.uu.se/](http://vald.astro.uu.se/) (given the terms of use for VALD3), but we provide instructions on how to do this in the `Cthulhu` documentation. `Cthulhu` currently supports ExoMol, HITRAN, HITEMP, and VALD line lists, though we welcome user requests for additional line list databases support. Once a line list has been downloaded, the user can move onto the next major use case of `Cthulhu`, computing cross sections.
 
