+++ joss/paper.md
@@ -41,23 +41,24 @@ Cross sections are typically pre-computed on a grid of pressures and temperature
 
 # Statement of Need
 
-JWST has recently significantly expanded the number of exoplanet and brown dwarfs with high-quality spectra spanning a wide wavelength range [e.g., @Beiler:2023; @Miles:2023; @Carter:2024; @Welbanks:2024]. High-fidelity spectra motivate detailed intercomparisons of exoplanet and brown dwarf modeling codes [e.g. @Barstow:2020], which often identify opacity database differences as a key modelling limitation. Ground-based high spectral resolution datasets [e.g., @Birkby:2017; @Snellen:2010; @Pelletier:2023] also critically rely on up-to-date opacity data, since older inaccurate line lists can lead to non-detections of molecules via cross correlations [e.g. @Merritt:2020] and propagate into retrieved chemical abundances [e.g. @Brogi:2019]. However, despite the key need to continually refine exoplanet and brown dwarf models with the latest state-of-the-art opacities, the process of calculating molecular and atomic cross sections is a non-trivial task that is typically outside the speciality of many exoplanet and brown dwarf researchers.
+JWST has recently significantly expanded the number of exoplanet and brown dwarfs with high-quality spectra spanning a wide wavelength range [e.g., @Beiler:2023; @Miles:2023; @Carter:2024; @Welbanks:2024]. High-fidelity spectra motivate detailed intercomparisons of exoplanet and brown dwarf modelling codes [e.g. @Barstow:2020], which often identify opacity database differences as a key modelling limitation. Ground-based high spectral resolution datasets [e.g., @Birkby:2017; @Snellen:2010; @Pelletier:2023] also critically rely on up-to-date opacity data, since older inaccurate line lists can lead to non-detections of molecules via cross-correlation [e.g. @Merritt:2020] or propagate into retrieved chemical abundances [e.g. @Brogi:2019]. However, despite the key need to continually refine exoplanet and brown dwarf models with the latest state-of-the-art opacities, the process of calculating molecular and atomic cross sections is a non-trivial task that is typically outside the speciality of many exoplanet and brown dwarf researchers.
 
-We have built `Cthulhu` to provide a user-friendly tool for beginners to learn how to work with the most commonly used line list databases and to readily calculate molecular and atomic cross sections. There are other open source codes that can calculate cross sections, such as HELIOS-K [@Grimm:2015; @Grimm:2021] and ExoCross [@Yurchenko:2018], that offer impressive computational performance and are excellent tools for experts to calculate cross sections. However, HELIOS-K requires Nvidia GPUs to run while ExoCross is built in Fortran, both of which can pose accessibility issues for beginners. We offer `Cthulhu`, a fully Python code designed to run on CPUs, as a user-friendly entry point into the world of cross sections for substellar atmospheres.
+We have built `Cthulhu` to provide a user-friendly tool for beginners to learn how to work with the most commonly used line list databases and to readily calculate molecular and atomic cross sections. There are other open source codes that can calculate cross sections, such as HELIOS-K [@Grimm:2015; @Grimm:2021] and ExoCross [@Yurchenko:2018], both of which offer impressive computational performance and are excellent tools for experts to calculate cross sections. However, HELIOS-K requires Nvidia GPUs to run while ExoCross is built in Fortran, both of which can pose accessibility issues for beginners. We offer `Cthulhu`, a fully Python code designed to run on CPUs, as a user-friendly entry point into the world of cross sections for substellar atmospheres.
 
 # Computing Molecular and Atomic Cross Sections with `Cthulhu`
 
[…]
