### 2024-08-23 #6894 @arnav-agrawal
@arjunsavel @LorenzoMugnai

Thank you both for your incredibly insightful comments which have greatly improved our paper and documentation. We have now fully implemented your suggestions. We also note that we have made several significant updates to our documentation:

- New code name and logo.
- We added a new ‘quick start’ tutorial, which covers making a full cross section for AlO and also packaging an opacity database as a HDF5 file. This tutorial is better placed to quickly highlight all the major features in Cthulhu.
- We have significantly improved and expanded on all the tutorials.

Below, we list how and where we addressed each of your comments in the paper.

@LorenzoMugnai

> You start with fundamental assumptions in the field (e.g., "Without accurate cross sections, one cannot reliably measure the chemical composition of substellar atmospheres."), which is a nice way to introduce the framework. However, I feel the paper could benefit from additional references or a few more lines to support such statements.

- We have now expanded the introduction to include several additional citations to the exoplanet and brown dwarf literature providing context.

> You mention that the code can run on a CPU laptop for "smaller line lists." What do you mean by "smaller"? Can you provide benchmarks, such as reference computation times and memory usage?

- We agree that this is an important detail. We find that 100 million transitions is about the limit for running quickly on a CPU. We have now added this number at the end of the Summary paragraph (and in a few places in the documentation). We also note that our expanded tutorial ‘Running Cthulhu on a Cluster’ covers how to use the code for significantly larger line lists.

> You state that the HDF5 output can be used with retrieval codes. Could you provide examples of codes natively supported by your output format (such as Poseidon, Taurex, etc.)? Or does the user need to adapt your output for these codes? Including references or citations for some retrieval codes would be helpful.

- Currently, the output HDF5 files produced by Cthulhu are in the format natively used by POSEIDON. We are planning to support the formats used by other radiative transfer codes shortly. We have clarified this in the paper.

> You graphically show that the cross sections produced can be used to generate transmission spectra. Can you explain how this can be done for a non-expert user, perhaps by referencing the integration with modelling codes?

- We agree that this is an important for non-experts to know. We now provide a brief description of how cross sections relate to radiative transfer in the last paragraph before the Future Developments.

> I feel the statement of need could be more effective if moved just after the summary, before delving into the details of Excalibur's computational process. Have you considered this?

- We agree. The Statement of Need has now been moved earlier in the paper.

> You close the manuscript by inviting the reader to provide suggestions. It would be useful to reference your "contribution guidelines," encouraging users to join the discussion on a dedicated channel (perhaps GitHub?).

- The Future Developments paragraph now addresses how users can contribute to the codebase via the “contribution guidelines”.

@arjunsavel

> Rapidly-evolving —> Rapidly evolving

- Fixed.

> Could even motivate opacity’s importance more by citing how sensitive at high-resolution spectroscopy can be to line list differences (e.g., Brogi and Line 2019).

- This is a very important point. We have added several citations to the Statement of Need paragraph on how inaccurate cross sections can affect results from high-resolution spectroscopy.

> Don’t need ELTs in parentheses, because the phrase is not used again.

- Fixed.

> Cite JWST and some of the ELT papers?

- We have added citations for each facility to the introduction.

> Please cite a few theory papers re: requiring cross-sections.

- Two citations to relevant exoplanet textbooks describing cross sections have been added. We also note that we have completely reworked the ‘introduction’ page in our documentation to describe cross sections in greater depth.

> Is it possible to make the legend text larger in Figure 1?

- We agree that the legends were too small, so the Figure has now been updated.

> Perhaps one sentence about what Excalibur does under the hood to calculate the cross-sections would be helpful.

- This is a very important point! We have added a sentence and reference to the thesis describing the algorithm underlying Cthulhu (MacDonald 2019). It’s a quite fun mathematical trick using complex derivatives ;)

> A HDF5 file --> an HDF5 file

- Fixed.

> A few more references would be helpful in laying out the radiative transfer. Cite the slant optical depth, for instance one of the earlier Fortney papers?

- Agreed, we have added Fortney (2005) to the end of Cthulhu paragraph and tweaked the wording to be clearer. We also now mention how cross sections are also applicable to other observing geometries.

> Please the higher-quality spectra of JWST.

- Added to Statement of Need.

> higher fidelity --> higher-fidelity

- Fixed.

> Please cite some brown dwarf modeling codes.

- Several brown dwarf modelling codes are now mentioned.

> Please cite some high-resolution ground-based datasets

- Citations have been added for several high-res data analyses.

> Beginner’s —> beginners

- Fixed.

> Speed-dependent —> speed-dependent

- Fixed.

Thanks again to everyone!
