### 2024-09-11 #6894 @LorenzoMugnai
@dfm and @arnav-agrawal,

I've reviewed both the paper and the documentation, and I believe they are now well-polished and ready for publication. I also tested the software, and it functions as described in the documentation (I especially enjoyed the "summon" method—great touch!). Congratulations to the authors for delivering such a well-packaged and useful tool.

I do have one minor suggestion, though it's more of a personal opinion. In the "Note" section of the contributing guidelines, I would suggest removing the endorsement of VS Code. While I agree that VS Code is an excellent tool (I use it myself and often recommend it to students and new Python users), I don't believe a specific recommendation belongs in the documentation. Other powerful IDEs, such as PyCharm, or emerging ones like Zed, offer similar capabilities. That said, the link to the Git branches guide is a valuable resource and should definitely remain.

On another note, I really liked the Git branch diagram in the contributing section. It's great to see this becoming a common feature in documentation (for reference, another example can be found in my own code's documentation: [ExoSim2 Git guidelines](https://exosim2-public.readthedocs.io/en/latest/contributing/guidelines.html#source-control)). However, I think there might be a minor typo in your diagram. Should the bug fix tag be labelled as v1.0.1 instead of v1.01?

Once again, thank you for your excellent work!
