### 2024-09-23 #6894 @arjunsavel
Thanks for pinging, and thanks for making all of these changes! This is a really excellent package, and I'm excited to start using it in my research. There are no barriers to acceptance on my end (@dfm).
