### 2024-09-29 #6894 @dfm
@arnav-agrawal — Thanks for all your work responding to the suggestions from the reviewers!

One last small change: Can you remove the code back ticks from around `Cthulhu` in the title of the manuscript? Those don't render properly after acceptance. Otherwise, I think we're ready to continue with the publication process.

Once you've fixed that:
