+++ setup.py
@@ -1,25 +1,19 @@
-import setuptools
+from setuptools import setup, find_packages
+import os
 
-with open("README.rst", "r") as fh:
-    long_description = fh.read()
-
-setuptools.setup(
-    name="Cthulhu", # Replace with your own username
-    version="1.0.0",
-    author="Ryan MacDonald, Arnav Agrawal",
-    author_email="ryanjmac@umich.edu, aa687@cornell.edu",
-    description="A python package to calculate atomic and molecular cross sections for substellar atmospheres.",
-
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    url="https://github.com/MartianColonist/Cthulhu",
-    packages=setuptools.find_packages(),
-    classifiers=[
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-    ],
-    python_requires='>=3.7.6',
+setup(
+    name = "Cthulhu", # Replace with your own username
+    version = "1.0.0",
+    description = "A python package to calculate atomic and molecular cross sections for substellar atmospheres.",
+    long_description = open(os.path.join(
+                            os.path.dirname(__file__), 'README.rst')).read(),
+    long_description_content_type = 'text/x-rst',
+    author = "Ryan MacDonald, Arnav Agrawal",
+    author_email = "ryanjmac@umich.edu, aa687@cornell.edu",
+    license = 'BSD 3-Clause License',
+    packages = ['Cthulhu'],
+    include_package_data = True,
+    python_requires = '>=3.7.6',
     install_requires = ['numpy',
                         'scipy',
                         'matplotlib',
@@ -32,5 +26,5 @@ setuptools.setup(
                         'lxml',
                         'hitran-api',
                         'pytest'],
-    zip_safe=False
+    zip_safe = False,
 )
