### 2024-10-03 a826f40
Update JOSS paper action
