+++ .github/workflows/draft-pdf.yml
@@ -1,11 +1,9 @@
-name: "Build Cthulhu Paper Draft"
+name: "Build JOSS Paper Draft"
 
 on:
   push:
     branches:    
-      - 'joss'
       - 'main'
-      - 'dev'
 
 jobs:
   paper:
@@ -13,7 +11,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -21,7 +19,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: joss/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
