+++ joss/paper.md
@@ -58,7 +58,7 @@ The first use of `Cthulhu` is to download existing molecular line lists from onl
 The foremost feature of `Cthulhu` is its ability to straightforwardly compute atomic and molecular cross sections at high speeds (typically > 100,000 lines per second on one CPU). `Cthulhu` calculates cross sections via a generalisation of the Vectorised Voigt method [@Yurchenko:2018], whereby our update to the algorithm uses complex derivatives to perturb a grid of pre-computed template Voigt profiles to the specific properties of each given line [see @MacDonald:2019, Chapter 5, for the mathematical description]. `Cthulhu` is accessible, as it does not require GPUs, can run on a standard laptop, and as a fully Python code it is easy for beginners to install and use. To compute a cross section, a user simply calls `Cthulhu`'s `compute_cross_section` function, specifying the location of the line list, the temperature and pressure, and the wavenumber range. More advanced users can specify custom settings via optional arguments (e.g. Voigt wing cutoffs, intensity cutoffs, or a user-provided pressure broadening file). The documentation and function docstrings explain the various arguments users can provide to `compute_cross_section`. The computed cross section is output by default as a .txt file in the `output` folder on the user's machine, which can be readily used to create an opacity database in a user-preferred format for a specific radiative transfer code. `Cthulhu` also offers utility functions to combine multiple cross section .txt files (e.g. a grid of cross sections for different temperature and pressures for one or more chemical species) into an HDF5 cross section database, as illustrated in our [`quick start`](https://cthulhu.readthedocs.io/en/latest/content/quick_start.html) guide.
 
 
-The cross section database HDF5 files produced by `Cthulhu` can be readily adapted for the user's favourite exoplanet or brown dwarf modelling or retrieval code. Cross section HDF5 files calculated by `Cthulhu` are already being used in the `POSEIDON` retrieval code [@MacDonald:2023] and the opacity formats for other retrieval codes will be natively supported soon. The lower part of \autoref{fig:Cthulhu_architecture} illustrates an application of `Cthulhu`'s cross sections, namely the calculation of exoplanet transmission spectra. Exoplanet transmission spectra modelling codes use cross sections to solve the equation of radiative transfer [e.g. @MacDonald:2022] by calculating the slant optical depth (the integral of the extinction coefficient along the line of sight, e.g. [@Fortney:2005]) for many rays of light passing through the exoplanet atmosphere. The final output of such a transmission spectrum calculation is the effective planet-star area ratio (i.e. transit depth) as a function of wavelength seen by a distant observer. Cross sections are similarly used in other observing geometries (e.g. secondary eclipses of exoplanets or directly imaged brown dwarfs) to calculate the attenuation of light along the path of each light ray.
[…]
