### 2024-10-04 #6894 @MartianColonist
@LorenzoMugnai, we implemented your final suggestions (a typo was fixed in our development flowchart, and we removed the specific VS Code endorsement from our contributions page). Thanks for your eagle-eye!

@dfm, we've incremented the version to v1.0.0 and confirm that @arnav-agrawal and I are happy with the manuscript.

The archived software for Cthulhu can be found on Zenodo at: https://doi.org/10.5281/zenodo.13888354

Thanks everyone, @LorenzoMugnai @arjunsavel, @dfm for all your support through this rewarding JOSS review! @arnav-agrawal and I feel Cthulhu is finally ready to be summoned 🐙
