### 2024-10-05 #6894 @dfm
@LorenzoMugnai, @arjunsavel — Many thanks again for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

@arnav-agrawal — I have now handed this off to @warrickball, the Associate Editor-in-Chief, for final approval. He may have some comments or edits, and then he'll start the publication process.
