### 2024-10-05 #6894 @warrickball
Hi @arnav-agrawal! :wave: Everything looks ready to go aside from one minor comment. The paper has at least three occurrences of "fully Python", which should be changed to "pure Python" (as in "a pure Python code" rather than "a fully Python code").

Though "fully Python" does seem to appear online, it would be unprecedented in JOSS, which so far only contains "pure Python".
