### 2024-10-05 8fc55b7
fully Python-> pure Python
