### 2024-10-05 #6894 @MartianColonist
Hi @warrickball, we've updated 'fully' -> 'pure' in the paper.
