### 2024-10-05 #6894 @warrickball
Thanks once again to @LorenzoMugnai and @arjunsavel for reviewing! JOSS simply wouldn't be possible without our volunteer reviewers.

Congratulations @arnav-agrawal and @MartianColonist, your paper has been published in JOSS!
